requirements.txt
pre-commit==3.6.0
black==23.12.1
beautifulsoup4
faiss-cpu
flake8==6.1.0
gigachat==0.1.36
python-dotenv
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.25
sentence-transformers==3.3.1
wikipediaapi